pyproject.toml
[project]
name = "SingleStopCoffea"
version = "0.5.0"
description = "A modular columnar analysis framework."
readme = "README.md"
requires-python = ">=3.12,<3.13"

dependencies = [
    "attrs>=25.4.0",
    "bokeh>=3.8.1",
    "cattrs>=25.3.0",
    "click>=8.3.0",
    "cloudpickle>=3.1.2",
    "coffea[xrootd,rucio]>=2025.10.0",
    "dask>=2025.3.0",
    "diskcache>=5.6.3",
    "distributed>=2025.3.0",
    "fsspec-xrootd>=0.5.1",
    "hist>=2.9.0",
    "histoprint>=2.6.0",
    "htcondor~=24.0.0",
    "ipython>=9.7.0",
    "jinja2>=3.1.6",
    "joblib>=1.5.2",
    "jupyter>=1.1.1",
    "lpcjobqueue",
    "lz4>=4.4.5",
    "mplhep>=0.4.1",
    "numpy>=2.3.4",
    "pyyaml~=6.0.0",
    "rich>=14.2.0",
    "scikit-learn>=1.7.2",
    "textual>=6.7.1",
    "textual-plotext>=1.0.1",
    "uv>=0.9.11",
]

[dependency-groups]
dev = [
    "ipykernel>=7.1.0",
    "pydata-sphinx-theme>=0.15.4",
    "pytest>=9.0.1",
    "sphinx>=8.2.3",
]

[tool.uv.sources]
lpcjobqueue = { git = "https://github.com/CoffeaTeam/lpcjobqueue.git" }
